Refactor RabbitMQ management client and permissions package
