+++ pkg/rabbitmq/management/permissions/types.go
@@ -1,0 +1,17 @@
+package permissions
+
+type permissionRequest struct {
+	User      string `json:"user"`
+	Vhost     string `json:"vhost"`
+	Configure string `json:"configure"`
+	Write     string `json:"write"`
+	Read      string `json:"read"`
+}
+
+type permissionsResponse []struct {
+	User      string `json:"user"`
+	Vhost     string `json:"vhost"`
+	Configure string `json:"configure"`
+	Write     string `json:"write"`
+	Read      string `json:"read"`
+}
